+++ Assets/Scripts/PlantCards/MagnifyingGrass.cs
@@ -8,8 +8,8 @@ public class MagnifyingGrass : Card
 	protected override IEnumerator OnThisPlay()
 	{
         yield return Glow();
-        if (GameManager.Instance.team == team) ChangeStats(GameManager.Instance.remainingTop, 0);
-        else ChangeStats(GameManager.Instance.opponentRemainingTop, 0);
+        if (GameManager.Instance.team == team) ChangeStats((int)GameManager.Instance.remainingTop, 0);
+        else ChangeStats((int)GameManager.Instance.opponentRemainingTop, 0);
 		yield return base.OnThisPlay();
 	}
